Added bottom navigation bar in finance app
